Move services from util folder
