+++ src/server/inversify.config.ts
@@ -3,13 +3,13 @@ import AWS from 'aws-sdk'
 import { container } from './util/inversify'
 import { GaLogger } from './api/analytics/analyticsLogger'
 import { DependencyIds } from './constants'
-import { CookieArrayReducer } from './util/transition-page'
+import { CookieArrayReducer } from './services/transition-page'
 import { OtpCacheRedis } from './api/cache/otp'
-import { MailerNode } from './util/email'
-import { CryptographyBcrypt } from './util/cryptography'
+import { MailerNode } from './services/email'
+import { CryptographyBcrypt } from './services/cryptography'
 import { DEV_ENV, accessEndpoint, bucketEndpoint, s3Bucket } from './config'
-import { MailerNoOp } from './util/emaildev'
-import { S3ServerSide } from './util/aws'
+import { MailerNoOp } from './services/emaildev'
+import { S3ServerSide } from './services/aws'
 import { UrlRepository } from './repositories/UrlRepository'
 import { UserRepository } from './repositories/UserRepository'
 import { UrlMapper } from './mappers/UrlMapper'
